rag/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==1.30.1
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
langchain-pinecone==0.0.1
pinecone-client==4.1.0
pymongo==4.7.2
motor==3.4.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.3
pydantic-settings==2.3.0
pypdf==4.2.0
python-docx==1.1.2
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.3.0
requests==2.32.3
beautifulsoup4==4.12.3
tiktoken==0.7.0
tenacity==8.3.0
aiofiles==23.2.1